ved_calendar/requirements.txt
# Core dependencies
aiogram>=3.4.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Database
aiosqlite>=0.19.0
psycopg2-binary>=2.9.9  # Для PostgreSQL (опционально)

# Templates
jinja2>=3.1.3

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# HTTP client
aiohttp>=3.9.1
httpx>=0.26.0

# Utilities
python-dateutil>=2.8.2
icalendar>=5.0.11

# Development
python-dotenv>=1.0.0
